chore: rename folder from repository to repositories
